Make the Z80 use 16 bits instead of 8 bit in IN and OUT

+++ src/applications/cpm_z80/cpm_application_session.cpp
@@ -42,7 +42,7 @@ void CpmApplicationSession::stop()
     throw std::runtime_error("Stop is not implemented for CP/M programs");
 }
 
-void CpmApplicationSession::out_changed(u8 port)
+void CpmApplicationSession::out_changed(u16 port)
 {
     if (port == s_finished_port) {
         m_is_finished = true;
