+++ src/chips/z80/interfaces/in_observer.h
@@ -8,6 +8,6 @@ class InObserver {
 public:
     virtual ~InObserver() = default;
 
-    virtual void in_requested(u8 port) = 0;
+    virtual void in_requested(u16 port) = 0;
 };
 }
